+++ packages/channels/src/event/PanChangeEvent.ts
@@ -1,8 +1,7 @@
 import { createEventClass } from 'seng-event';
-import { HasSignalModifier } from '../types';
 
 export type PanChangeData = {
-  target: HasSignalModifier;
+  pan: number;
 };
 
 export class PanChangeEvent extends createEventClass<PanChangeData>()(