+++ packages/channels/src/testing/testUtils.ts
@@ -16,9 +16,6 @@ export const mockXMLHttpRequest = () => {
 };
 
 export const createMockChannelsInstance = (sounds?: Array<CreateSound>) => {
-  (window as any).WebAudioTestAPI.setState({
-    'AudioContext#createStereoPanner': 'enabled',
-  });
   const channelsInstance = new Channels({
     soundsPath: 'path',
     soundsExtension: 'mp3',