+++ packages/channels/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Volume/pan change events are now dispatched from where the change happened: a Channel, PlayingSound or the Channels instance (for main output) 
+
 ## [0.6.0] - 2022-09-07
 - Effects can be set pre or post volume (on a sound, a channel or main output)
 - Fix stopAllSounds() not immediately stopping sounds when fadeOutTime is set